// src/app/accounts/models/costcentert.ts
import { GlobalVariables } from '../../core/models/globalvariables';

export class CostCentert {

    ct_pkid: string;
    ct_year: number;
    ct_jvh_id: string;
    ct_jv_id: string;
    ct_acc_id: string;
    ct_category: string;
    ct_cost_id: string;
    ct_cost_code: string;
    ct_cost_name: string;
    ct_cost_year: number;
    ct_amount: number;
    ct_ctr: number;
}